Swap Rosenbrock23 for Rodas5 and use StaticArrays
Hey, someone cued me into this benchmark. Rosenbrock23 should only be used for high tolerance problems, so it's not the appropriate algorithm here. It's going to be slow at 1e-8 as the documentation mentions: a better fit would be Rodas5. And for a small problem like that, it's recommended that one uses static arrays. The final time I am seeing is 75μs, but my computer seems a little slower than yours so I would expect somewhat of a difference. Overall nice work but just gotta amend that!