Update DifferentialEquations.jl benchmarks
DifferentialEquations.jl went through a lot of changes since this. The biggest one is that we don't have a dumb BLAS anymore, instead we use a pure-Julia BLAS which does a lot better on small problems then OpenBLAS did (it had a lot of overhead), which brings the array code down to where it should've been. But the static array codes also got a lot more recent optimizations, so this updates the static array timings as well which now make a much strong case for static binding small sized solvers.